Move FragmentStrictMode back into separate package
